requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
Flask-MySQL==1.6.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
protobuf==4.21.12
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.31.0
urllib3==2.5.0
Werkzeug==3.1.3
